Exercise 4.21: Only user who create blog can delete it

+++ part4/blog-list/backend/controllers/blogs.js
@@ -64,8 +64,22 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  console.log('decoded Token: ', decodedToken);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+
+  if (decodedToken.id === blog.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    return response.status(204).end();
+  } else {
+    return response
+      .status(401)
+      .json({ error: 'You do not have permission to delete blog post!' });
+  }
 });
 
 module.exports = blogsRouter;
